feat(game): The Ship.prototype.update Method
push the ship in the direction it’s pointing using the correct power, but only if this.thruster_on is true
we need to use function.apply to call the Mass.prototype.update method, passing in our this keyword to reference the ship instance and the arguments list passed into the function. This is a generic way to call methods of a parent class when the child class has overridden them.

#58