feat(game): Hook Up Projectiles in the update Function
We update the projectiles as expected in a forEach loop over the array. Then we check the value of projectile.life. If the value is 0 or less, we call the array. prototype.splice method on our projectiles array to remove the dead projectile. This removes all references to the projectile, and it gets removed from memory.
After the loop, we check a new property: ship.trigger. If it’s set, we push a new projectile onto the projectiles array. We get the projectile by calling ship.projectile and passing in the elapsed time.

#60