refactor(game): The AsteroidsGame Constructor
The constructor takes one argument: the id attribute of the canvas element we want to draw our game onto
We use this to create a reference to a canvas.
We also extract a reference to the 2D canvas context and store it as this.c for easy access.
We then set the focus on the canvas so key presses will register properly
We then set some properties that were previously either global variables or were “magic numbers” that appeared in the code
Then we create a new ship and set the projectiles and asteroids properties to blank arrays
We add a single moving asteroid into the asteroids array by calling the new AsteroidsGame.prototype.moving_ asteroid helper method
Finally, we connect some event handler code and start the game loop
we “bind” these callback functions to the game instance